refactor(homepage): decompose index.vue into focused components
Extract 727-line monolithic page into 6 focused components under
app/components/homepage/ with shared types in app/types/homepage.ts.
No behavioral, style, or animation changes.

- app/types/homepage.ts: centralizes 8 inline TS types
- HomepageHero.vue: hero section + logos + heroTitle computed
- HomepageMetrics.vue: metrics stats grid
- HomepageImmersiveStage.vue: immersive 4-section wrapper (Problem,
  Guide, Mechanism, Process) + scroll composable
- HomepageFeatures.vue: 3-col feature card grid
- HomepageAbout.vue: profile photo/quote/bio/milestones
- HomepageCta.vue: CTA with GradientGlow
- index.vue: slimmed to ~57-line orchestration layer
